feat:add author handler

+++ author_management/internal/entity/author.go
@@ -1,7 +1,7 @@
 package entity
 
 type Author struct {
-	Id       uint
+	Id       uint64
 	Name     string
 	PhotoUrl string
 	Gender   string
